chore: moved chain id logic to Chain class
