+++ src/helpers/index.ts
@@ -5,4 +5,3 @@ export * from './not-empty'
 export * from './number-to-hex'
 export * from './parse-custom-error'
 export * from './sleep'
-export * from './chain'
